refactor: 공통적으로 사용되는 상수들을 따로 클래스로 관리하지 않고 pulbic으로 open해서 사용하도록 변경
