--- src/main/java/lotto/domain/lotto/utils/LottoAttributes.java
@@ -1,10 +1,0 @@
-package lotto.domain.lotto.utils;
-
-public final class LottoAttributes {
-
-    public static final int LOTTO_NUMBER_START = 1;
-    public static final int LOTTO_NUMBER_END = 45;
-    public static final int LOTTO_LINE_PRICE = 1000;
-    public static final int LOTTO_NUMBER_COUNT = 6;
-
-}
